komentarze

+++ src/main/java/com/meethub/domain/model/response/MonthlyTrend.java
@@ -1,13 +1,15 @@
+// MonthlyTrend.java - ZMIEŃ z Double na BigDecimal
 package com.meethub.domain.model.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;  // DODAJ
 
 @Data @Builder @NoArgsConstructor @AllArgsConstructor
 public class MonthlyTrend {
     private String monthName;
     private Integer meetingsCount;
-    private Double avgAttendance;
+    private BigDecimal avgAttendance;  // ZMIENIONE z Double
 }
